requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
openai==1.3.7
python-multipart==0.0.6
google-cloud-vision==3.4.4
beautifulsoup4==4.12.2
fake-useragent==1.4.0
